chore: Sprite 크기 조정

+++ supermaker_assignment/Assets/Scripts/Unit/UnitMetaData.cs
@@ -13,6 +13,6 @@ namespace Unit
         public EUnitType Type;
         public uint SellPrice;
         public Sprite Sprite;
-        public uint ScaleSize;
+        public float ScaleSize;
     }
 }
